: PEP8 and fixed imports

+++ illum/point_source_to_zon.py
@@ -1,7 +1,8 @@
 #!/usr/bin/env python3
 
 import numpy as np
-import pyproj, yaml
+import pyproj
+import yaml
 from collections import defaultdict as ddict
 
 # Read domain.ini
@@ -14,16 +15,16 @@ with open(dom_name) as f:
 #    lat lon pow hobs dobs fobs hlamp S_type P_type
 inv_name = input("    LOP inventory filename : ")
 
-inv = np.loadtxt(inv_name,dtype=object)
-inv[:,:7] = inv[:,:7].astype(float)
+inv = np.loadtxt(inv_name, dtype=object)
+inv[:, :7] = inv[:, :7].astype(float)
 
 # Classify point by pixel
 data = ddict(list)
 
 for light in inv:
-    lat,lon = light[:2]
+    lat, lon = light[:2]
 
-    p1 = pyproj.Proj("epsg:4326") # WGS84
+    p1 = pyproj.Proj("epsg:4326")  # WGS84
     p2 = pyproj.Proj(domain['srs'])
     x, y = pyproj.transform(p1, p2, lon, lat, always_xy=True)
 
@@ -37,20 +38,20 @@ for light in inv:
             row = int((y-extent['ymin'])/extent['pixel_size'] - 0.5)
             break
 
-    data[level,col,row].append(light[2:])
+    data[level, col, row].append(light[2:])
 
 # Generate zone inventory
 out_name = input("    Output filename : ")
-with open(out_name,'w') as inv_file:
+with open(out_name, 'w') as inv_file:
     inv_file.write("# X	Y	R	hobs	dobs	fobst   hlamp	Zone inventory\n")
-    for level,col,row in data:
-        lights = np.asarray(data[level,col,row])
-        pow_tot = np.sum(lights[:,0])
+    for level, col, row in data:
+        lights = np.asarray(data[level, col, row])
+        pow_tot = np.sum(lights[:, 0])
 
-        ho = np.average(lights[:,1],weights=lights[:,0])
-        do = np.average(lights[:,2],weights=lights[:,0])
-        fo = np.average(lights[:,3],weights=lights[:,0])
-        hl = np.average(lights[:,4],weights=lights[:,0])
+        ho = np.average(lights[:, 1], weights=lights[:, 0])
+        do = np.average(lights[:, 2], weights=lights[:, 0])
+        fo = np.average(lights[:, 3], weights=lights[:, 0])
+        hl = np.average(lights[:, 4], weights=lights[:, 0])
 
         frac = ddict(float)
         for i in range(len(lights)):
@@ -60,13 +61,13 @@ with open(out_name,'w') as inv_file:
         x = extent['pixel_size'] * (col - 0.5) + extent['xmin']
         y = extent['pixel_size'] * (row - 0.5) + extent['ymin']
 
-        lon,lat = pyproj.transform(p2,p1,x,y,always_xy=True)
+        lon, lat = pyproj.transform(p2, p1, x, y, always_xy=True)
 
         data_line = ("%.06f\t"*2+"%g\t"*5+"%s\n") % \
-                    ( lat, lon,
-                      extent['pixel_size']/2000., # diameter in km to radii in m
-                      ho, do, fo, hl,
-                      ' '.join( ["%g_%s_%s" %
-                                (frac[i],i[0],i[1]) for i in frac] ) )
+                    (lat, lon,
+                     extent['pixel_size']/2000.,  # diameter in km to radii in m
+                     ho, do, fo, hl,
+                     ' '.join(["%g_%s_%s" %
+                               (frac[i], i[0], i[1]) for i in frac]))
 
         inv_file.write(data_line)
